# laravelapi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "API documentation for your Laravel backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/payment": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Get all payments",
                "operationId": "318a8e6d1447a7cbd66237ccd6660678",
                "responses": {
                    "200": {
                        "description": "List of payments"
                    }
                }
            },
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Create a new payment",
                "operationId": "3bf87a1d2db6171310016de0615a4707",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "user_id"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "credit_card",
                                            "paypal",
                                            "bank_transfer"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "completed",
                                            "failed",
                                            "refunded"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment created"
                    }
                }
            }
        },
        "/api/payment/{id}": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Get a payment by ID",
                "operationId": "5ec820dc5a59156a714c50a583e10016",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment details"
                    },
                    "404": {
                        "description": "Payment not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Payment"
                ],
                "summary": "Delete a payment",
                "operationId": "3e64fcf6d186d746c69967dc5467c639",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment delete successful"
                    },
                    "404": {
                        "description": "Payment not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Update a payment by ID",
                "operationId": "5fd1263037169f9a22eb7d88d877b09a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "credit_card",
                                            "paypal",
                                            "bank_transfer"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "completed",
                                            "failed",
                                            "refunded"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment updated successfully"
                    },
                    "404": {
                        "description": "Payment not found"
                    }
                }
            }
        },
...
    },
...
}